requirements.txt
Django==1.11.7
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Markdown==2.6.11
psycopg2==2.7.3.2
PyJWT==1.5.3
pymongo==3.5.1
pytz==2017.3
